Add(minjage/refactor): 리팩토링 #113
