+++ app/pages/boards/device.js
@@ -1,13 +1,13 @@
 import React from "react";
 import Head from "next/head";
-import { useGetBoardList } from "../../hooks/useGetBoardList";
+import { useGetMarketList } from "../../hooks/useGetMarketList";
 
 import BoardBanner from "../../components/Board/BoardBanner";
 import Market from "../../components/Board/Market";
 
 const MarketDevicePage = () => {
   const categoryName = "device";
-  const { productList } = useGetBoardList(categoryName);
+  const { productList } = useGetMarketList(categoryName);
 
   return (
     <>
