+++ app/pages/boards/notice.js
@@ -1,41 +1,14 @@
-import React, { useState, useEffect, useRef } from "react";
-import axios from "axios";
+import React from "react";
 import BoardBanner from "../../components/Board/BoardBanner";
 import Basic from "../../components/Board/Basic";
 import Head from "next/head";
-import { useRouter } from "next/router";
+import { useGetBasicBoardList } from "../../hooks/useGetBasicBoardList";
+
+const categoryName = "notice";
 
 const BoardNoticePage = () => {
-  const [pageNumber, setPageNumber] = useState(0);
-  const [boardList, setBoardList] = useState([]);
-
-  const perPage = useRef(10);
-  const pageVisited = pageNumber * perPage.current;
-  const pageCount = Math.ceil(boardList.length / perPage.current);
-  const categoryName = "notice";
-
-  const router = useRouter();
-
-  useEffect(() => {
-    axios
-      .get(`/api/boards/${categoryName}`)
-      .then((response) => {
-        if (response.data.success) {
-          const result = response.data.boards;
-          setBoardList(result);
-        }
-      })
-      .catch((err) => {
-        const response = err.response;
-        if (response.status === 400) {
-          alert(response.data.msg);
-        }
-      });
-  }, [categoryName, router.pathname]);
-
-  const changePage = ({ selected }) => {
-    setPageNumber(selected);
-  };
+  const { boardList, perPage, pageVisited, pageCount, changePage } =
+    useGetBasicBoardList(categoryName);
 
   return (
     <>
